+++ src/fides/api/ops/models/policy.py
@@ -37,8 +37,7 @@ class CurrentStep(EnumType):
     access = "access"
     erasure = "erasure"
     consent = "consent"
-    erasure_email_post_send = "erasure_email_post_send"
-    consent_email_post_send = "consent_email_post_send"
+    email_post_send = "email_post_send"
     post_webhooks = "post_webhooks"